requirements.txt
Flask==2.0.1
flask-restful==0.3.9
flask-mongoengine==1.0.0
flask-bcrypt==0.7.1
marshmallow==3.14.1
python-dotenv==0.19.2